Added modularized code structure for improved organization and maintenance
