+++ src/modules/createRow.js
@@ -1,0 +1,7 @@
+const createRow = () => {
+  const row = document.createElement('div');
+  row.classList.add('row');
+  return row;
+};
+
+export default createRow;
